문의 파일통합 파일 매니저 사용, 문의 리뷰, 사진 없을 때도 등록, 수정, 삭제 가능하게 구현

+++ peachStore/src/main/java/peachstore/service/inquiry/InquiryServiceImpl.java
@@ -1,6 +1,7 @@
 package peachstore.service.inquiry;
 
 import java.time.LocalDateTime;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -19,29 +20,28 @@ import peachstore.exception.Uploadexception;
 import peachstore.repository.Inquiry.InquiryDAO;
 import peachstore.repository.Inquiry.InquiryImgDAO;
 import peachstore.service.admin.AdminService;
+import peachstore.util.FileCommonManager;
 import peachstore.util.FileManager;
 
 /**
  * 문의하기 service
- * @author 성유진 
+ * 
+ * @author 성유진
  * @since 2025-07-26
  */
 @Slf4j
 @Service
-public class InquiryServiceImpl implements InquiryService{
+public class InquiryServiceImpl implements InquiryService {
 
 	@Autowired
 	private InquiryDAO inquiryDAO;
 	@Autowired
 	private InquiryImgDAO inquiryImgDAO;
 	@Autowired
-	private FileManager fileManager;
-	
-//	@Autowired
-//	private InquiryImgService inquiryImgService;
+	private FileCommonManager filaCommonManager;
+
 	@Autowired
 	private AdminService adminService;
-
 
 	@Override
 	public List selectAll(Inquiry inquiry) {
@@ -55,47 +55,77 @@ public class InquiryServiceImpl implements InquiryService{
 
 	@Transactional
 	@Override
-	public void regist(Inquiry inquiry, String savePath) throws InquiryException, InquiryImgException, Uploadexception{
+	public void regist(Inquiry inquiry, String savePath) throws InquiryException, InquiryImgException, Uploadexception {
 		inquiryDAO.insert(inquiry);
-		
-			boolean flag = fileManager.save(inquiry, savePath);
-			log.debug("filemanager.save result: {}", flag);
-			if (flag) {
-				log.debug("if문 진입");
-				List<InquiryImg> imgList = inquiry.getImgList();
-				for(InquiryImg inquiryImg : imgList) {
-					inquiryImg.setInquiry(inquiry);
-					inquiryImgDAO.insert(inquiryImg);
-				}
-			}
+
+		// 디렉토리 생성
+		String subDir = "p_" + inquiry.getInquiry_id();
+
+		// 이미지 저장
+		List<String> savedFilenames = filaCommonManager.saveFiles(inquiry.getPhoto(), savePath, subDir);
+
+		List<InquiryImg> imgList = new ArrayList<>();
+		for (String filename : savedFilenames) {
+			InquiryImg inquiryImg = new InquiryImg();
+
+			inquiryImg.setInquiry(inquiry);
+			inquiryImg.setFilename(filename);
+			inquiryImgDAO.insert(inquiryImg);
+
+			imgList.add(inquiryImg);
+		}
+
+		inquiry.setImgList(imgList);
 	}
 
 	@Transactional
 	@Override
 	public void update(Inquiry inquiry, String savePath) throws InquiryException {
-		
-		//기존 레코드 삭제
-		inquiryImgDAO.delete(inquiry.getInquiry_id());
-		
-		//폴더 안의 기존 파일 삭제
-		fileManager.remove(inquiry, savePath);
-		
-		inquiryDAO.update(inquiry);
-		
-		//새로 저장 
-		fileManager.save(inquiry, savePath);
-		
-		List<InquiryImg> imgList = inquiry.getImgList();
-		for(InquiryImg inquiryImg : imgList) {
-			inquiryImg.setInquiry(inquiry);
-			inquiryImgDAO.insert(inquiryImg);
+		List<InquiryImg> img = inquiryImgDAO.select(inquiry.getInquiry_id());
+
+		if (img != null && !img.isEmpty()) {
+			// 기존 이미지 삭제
+			inquiryImgDAO.delete(inquiry.getInquiry_id());
 		}
+
+		// 폴더 안의 기존 파일 삭제
+		String subDir = "p_" + inquiry.getInquiry_id();
+		filaCommonManager.remove(subDir, savePath);
+
+		// 문의 내용 수정
+		inquiryDAO.update(inquiry);
+
+		// 새로 저장
+		List<String> savedFilenames = filaCommonManager.saveFiles(inquiry.getPhoto(), savePath, subDir);
+
+		List<InquiryImg> imgList = new ArrayList<>();
+		for (String filename : savedFilenames) {
+			InquiryImg inquiryImg = new InquiryImg();
+
+			inquiryImg.setInquiry(inquiry);
+			inquiryImg.setFilename(filename);
+			inquiryImgDAO.insert(inquiryImg);
+
+			imgList.add(inquiryImg);
+		}
+
+		inquiry.setImgList(imgList);
 	}
-	
+
 	@Override
 	public void remove(Inquiry inquiry, String savePath) throws InquiryImgException, InquiryException {
-		inquiryImgDAO.delete(inquiry.getInquiry_id());
-		fileManager.remove(inquiry, savePath);
+
+		List<InquiryImg> img = inquiryImgDAO.select(inquiry.getInquiry_id());
+
+		if (img != null && !img.isEmpty()) {
+			// 기존 이미지 삭제
+			inquiryImgDAO.delete(inquiry.getInquiry_id());
+		}
+
+		// 폴더 안의 기존 파일 삭제
+		String subDir = "p_" + inquiry.getInquiry_id();
+		filaCommonManager.remove(subDir, savePath);
+
 		inquiryDAO.delete(inquiry);
 	}
 
@@ -108,7 +138,7 @@ public class InquiryServiceImpl implements InquiryService{
 	 * 문의에 답변 달기
 	 */
 	@Override
-	public void answerInquiry(int inquiry_id, int admin_id, String answer_text) throws InquiryException{
+	public void answerInquiry(int inquiry_id, int admin_id, String answer_text) throws InquiryException {
 		// 해당 문의 조회
 		Inquiry inquiry = inquiryDAO.selectById(inquiry_id);
 
@@ -116,13 +146,14 @@ public class InquiryServiceImpl implements InquiryService{
 		Admin admin = adminService.findById(admin_id);
 		log.debug("admin 객체: {}", admin);
 		log.debug("inquiry.setAdmin 전 상태: {}", inquiry);
-		
+
 		// 답변 정보 추가
 		inquiry.setAdmin(admin);
 		inquiry.setAnswer_text(answer_text);
 		inquiry.setAnswered_at(LocalDateTime.now());
-		log.debug("anwerInquiry data - {}, {}, {}", inquiry.getAdmin().getAdminId(), inquiry.getAnswer_text(), inquiry.getAnswered_at());
-		
+		log.debug("anwerInquiry data - {}, {}, {}", inquiry.getAdmin().getAdminId(), inquiry.getAnswer_text(),
+				inquiry.getAnswered_at());
+
 		// 데이터베이스 업데이트
 		int result = inquiryDAO.updateAnswer(inquiry);
 		if (result == 0) {
@@ -134,14 +165,12 @@ public class InquiryServiceImpl implements InquiryService{
 	 * pk로 문의 조회
 	 */
 	@Override
-	public Inquiry findById(int inquiry_id) throws InquiryException{
+	public Inquiry findById(int inquiry_id) throws InquiryException {
 		Inquiry inquiry = inquiryDAO.selectById(inquiry_id);
 		if (inquiry == null) {
 			throw new InquiryException("문의가 존재하지 않습니다.");
 		}
 		return inquiry;
 	}
-
-	
 
 }
